Enough work for today
The TMSiGUI currently does work with psychtoolbox. There is still no direct signal sent between the two of them but I am getting close. prompter1function2 is an experimental code that should use the global variable rec which is set by the TMSi. It isn't working yet but prompter1function and TMSiGUI work