Made some changes - ugly workarounds used
I'm not proud of this one, it uses some pretty ugly code but it should reliable enough to actually use.  The save button for TMSI will pause until the F8 button is pressed. The same goes for the opening prompt with psychtoolbox. Having them both activated by the same key press should sync the two