Added Poly5 file management
Due to the difficulties in getting the TMSi function to do what I want
it to do, I have created code that moves each block of data to its
matching folder. A check is performed at the start of each block. If an
EEG recording session exists, that file is renamed to the previous block
and moved. This should enable the save function to continue using the
same name for each output file. Main thing to be wary of is to check the
EEG file after we quit the program.