Fixed overwriting problem, added beeping
The code now beeps when it gets to the end of a session. The goal here
will be to get the EEG recording to save when you press F9 as well, this
way, they should be totally in sync. Overwriting problem was occurring
due to sloppy coding. Should be fixed now. Times have been readjusted,
so the code can be actually run again.