File saving revision - unproven with actual system
Made some changes to the timing file saving, the timing files are now
saved to the running folder. The rationnale for this is that we have a
post-session script which will combine all the data from the session and
move the raw files. Given the poly5 files need to be in the running
folder, it is easier to keep them all in the same directory. Please
ensure that you test this on the real system as it hasn't been tested
there yet.